Port async rediscache features to new package
This initial commit ports the async rediscache features that we wrote
for our bots to a standalone package. This makes it easier to maintain
the code as we do not have to maintain it in multiple locations.

This initial commit contains a minimal build&publish pipeline as well as
a Pipfile to make local development easier. It is not yet ready for
production.