Fix incorrect handling of bools in type converter
The `RedisObject._to_typestring` static method accidentally marked any
`bool` value it received as the first type in sequence of allowed types
it was iterating over.

This commit changes the logic by matching the value to the correct type
first and only then doing the bool-to-int conversation.

This commit fixes #2