Remove check for duplicate namespaces
Initially, we had a check for duplicate namespaces and warned the user
when they were trying to set the same namespace for a new or different
RedisObject instance. However, with the introduction of queuing systems,
where a producer and consumer both connect to the same queue, such a
protection stands in the way of the actual functionality.

As a result, I've removed the namespace check altogether, which means
that it's possible to create RedisObjects that share a namespace with
another RedisObject instance.