Make compound transactions atomic within namespace
As we provide more than just a thin wrapper around the operations Redis
offers, this means that some of our methods can be affected by race
conditions.

Take this operation, for instance:

```py
async def pop(self, key, default):
    value = await self.get(key, default)
    await self.delete(key)
    return value
```

As Redis does not offer a POP command for hash sets, pop functionality
can only be implemented by invoking two Redis commands: first, you get
the value and, second, you delete the key.

This means that if another coroutine inserts another value between those
two operations, we'll delete the newly set value, but return the old
value.

To prevent such race conditions, a namespace is now locked to completing
an entire operation, including compound operations, before initiating a
new operation.

As all operations needs to be locked, not only the compound operations,
I've introduced a new decorator, `namespace_lock`, that can be used on
all methods that need to be locked. If a locked method needs to calls
another locked method, it can pass the `acquire_lock=False` kwarg to
prevent a deadlock situation.

To make debugging the lock situation a bit easier, I've also created an
`asyncio.Lock` subclass, `NamespaceLock`, that is aware of the namespace
it's locking and can therefore display that information in its repr.

One limitation of atomifying compound operations like this is that the
operations are only atomic within an application. If another application
mutates items in the same namespace, this locking mechanism will be
bypassed.

Signed-off-by: Sebastiaan Zeeff <sebastiaan.zeeff@gmail.com>