Exclude tests package from PyPI release
Previously, we used `setuptools.find_packages` to automatically find
Python packages in the root directory of our PyPI package. While this is
an easy way of specifying the packages that should packaged into one
PyPI package, this meant that our `tests/` package got included as a
separate package.

This also meant that `pip` would install `tests/` as a separate package
to `site-packages` when installing our PyPI package.

As we only have a single package directory that needs to be included,
I've chosen to specify it directly instead of using `find_packages` with
a bunch of ignores.

Signed-off-by: Sebastiaan Zeeff <sebastiaan.zeeff@gmail.com>