Remove `dev` as optional feature of our release
We had a `dev` optional feature in our release that would install
`flake8` to create a minimal development environment. However, as we
also use a lot of flake8 plugins in our typical dev setup, I decided to
remove this option from `extras_require` and ask people to create an
environment using the `Pipfile` instead.

Signed-off-by: Sebastiaan Zeeff <sebastiaan.zeeff@gmail.com>