Reverse direction of RedisQueue for consistency
A new type that's going to be introduced, the RedisTaskQueue, requires
the use of two methods that always push an item to the left side of the
Redis list and pop an item from the right side of the Redis list.

However, our RedisQueue type did the opposite: it popped from the left
and pushed to the right side of the Redis List. While this behavior is
not directly exposed to the end-user, I've decided to reverse these
directions to make sure that our two Queue objects use the same pop/push
pattern.

Signed-off-by: Sebastiaan Zeeff <sebastiaan.zeeff@gmail.com>