Pop password kwarg when connection to fakeredis
Whenever `password` is not `None` while using `fakeredis`, a `AUTH`
command will be send to the `FakeSever` set up by `fakeredis`. However,
this fake server does not support the AUTH command, which causes it to
panic and return an error.

The solution is just to never provide a `password` kwarg when using
`fakeredis`, as it's not supported.

Signed-off-by: Sebastiaan Zeeff <sebastiaan.zeeff@gmail.com>