Add deprecation warning to NamespaceLock
Using a namespace lock to prevent race conditions between different
operations performed on the same Redis Cache is phased out of the
application. The main reasons are that it's fairly easy to end up in a
deadlock situation and that client-level locking does not prevent race
conditions in a multi-client setup.

Instead of giving the impression that a namespace lock would solve all
race condition problems, I've decided to phase out the lock by adding a
deprecation warning that signals removal in version 1.0.0.

As a replacement, I'm planning to make each built-in redis operation
atomic on the Redis-side by using Redis/Lua scripts for compound
operations that consist of multiple Redis commands.

Internally, the lock will still be used until version 1.0.0 to guarantee
that applications still using this feature will lock as expected. As
internal usage is triggered even when the downstream application does
not use such locks, the deprecation hidden won't be issued for internal
usage.

Signed-off-by: Sebastiaan Zeeff <sebastiaan.zeeff@gmail.com>