Allow loading/registering Lua scripts in Redis
To make some of our compound operations atomic from a Redis point of
view, instead of a per client point of view, we need to execute those
multiple Redis commands as a single command using a Lua/Redis script.

The feature added allows us to load a specific script registered with
Redis, falling back to registering it if it isn't found. The latter is
really important, as a Redis service may be restarted independent of a
client application.

As registering scripts is based on the SHA1 hash of the script,
different versions of RedisCache will use the version of the script they
expect, as they will select the appropriate script version by hash.

Signed-off-by: Sebastiaan Zeeff <sebastiaan.zeeff@gmail.com>