Make `increment` atomic from a redis perspective
While we guaranteed the atomicity of an increment/decrement operation
from a single client point-of-view, the operation was not atomic from a
Redis perspective: If multiple clients tried to modify the value, race
condition could still occur.

As an alternative, I've introduces a Redis Lua script that performs the
compound operation at the Redis side. As Redis scripts are executed
atomically, this makes the entire compound operation atomic.

Signed-off-by: Sebastiaan Zeeff <sebastiaan.zeeff@gmail.com>