Make `pop` method atomic from a Redis perspective
Similarly to making the increment method atomic, the `pop` method has
been made atomic by using a Redis script. The reason we need to do this
is because a `pop` operation consists of two Redis commands: We want to
HGET a value and then HDEL the key->value pair.

By performing these operations in a single Redis script, those two
commands are executed atomically.

Signed-off-by: Sebastiaan Zeeff <sebastiaan.zeeff@gmail.com>