Add rescheduling scripts for atomic rescheduling
Our RedisTaskQueue should be able to atomically reschedule a task that
was previously added to the task queue. While this normally shouldn't
happen, it may happen if a client aborts unexpectedly. It should also be
able to atomically reschedule all pending tasks for a given client, for
the same reasons.

The scripts I've added accomplish just that.

Signed-off-by: Sebastiaan Zeeff <sebastiaan.zeeff@gmail.com>