Add RedisTaskQueue type with task tracking feature
I've added a new datatype: You can now use a TaskQueue to safely process
tasks without having to worry about losing tasks if the client closes
unexpectedly.

When a task is consumed from the queue, the TaskQueue will move it into
a "pending" tasks queue where it will live until the task is marked as
done. If tasks will not be completed, they can be rescheduled in the
main queue, ready to be consumed another time.

In general, clients should reschedule all tasks in the "pending" tasks
queue on start-up: If there are still tasks left in that queue, they
apparently weren't processed fully during the previous execution of the
client. However, do note that it could be the case that the client has
processed the task partially.

In a multi-client setup, clients should create an instance with a unique
client ID. This means they will have their own pending task queue. If
clients then restart with their old ID, they can process their own
pending tasks queue.

Signed-off-by: Sebastiaan Zeeff <sebastiaan.zeeff@gmail.com>