Migrate To aioredis v2
Bumps the aioredis version to 2.x, and makes all the changes required to
adopt the new API. This includes bumping fakredis to a version which
supports aioredis 2.

A lot of the changes have to do with the removal of explicit pool
management. There are a few API changes, such as `evalsha`, as well as
the removal of some error types (not documented). Most of these changes
are unfortunately undocumented in the changelog and migration guides.

Refer to the following for a general overview of changes in v2:
https://aioredis.readthedocs.io/en/latest/migration/

Signed-off-by: Hassan Abouelela <hassan@hassanamr.com>