Switch To redis-py
Switches out aioredis with redis-py, which have been combined starting
from version 4.2.0. Since aioredis v2 is mostly an accurate subset of
redis-py, not much had to change. The most major change is that redis-py
does expose the connection pool unlike aioredis, which allows the
session to re-expose the pool property.

Signed-off-by: Hassan Abouelela <hassan@hassanamr.com>